fix(tools): PDF 压缩改用 zlib 后端别名并修复 dict 查询、拼图 overlay 坐标类型

+++ src-tauri/src/commands/tools/image.rs
@@ -323,7 +323,7 @@ pub async fn image_stitch(
             } else {
                 (0, offset)
             };
-            image::imageops::overlay(&mut canvas, &scaled, x, y);
+            image::imageops::overlay(&mut canvas, &scaled, i64::from(x), i64::from(y));
             offset += if is_horizontal {
                 scaled.width()
             } else {
